Fixed the pricing endpoint to work with new code

+++ model/PricingAdjustment/PricingAdjustment.go
@@ -4,8 +4,8 @@ type PricingAdjustment struct {
 	AdjustmentDescription string  `json:"adjustmentDescription"`
 	QuantityAdjusted      int     `json:"quantityAdjusted"`
 	PricingModel          string  `json:"pricingModel"`
-	TotalAdjustment       float32 `json:"totalAdjustment"`
-	Amount                float32 `json:"amount"`
+	TotalAdjustment       float64 `json:"totalAdjustment"`
+	Amount                float64 `json:"amount"`
 }
 
 func NewPricingAdjustment() *PricingAdjustment {
